[TC] No compile-time error on different method signatures
int m(int, float) and int m(float, int) does not raise an error anymore

+++ tests/typing/should_fail/same_args_in_method.java
@@ -1,7 +1,7 @@
 package tutu.titi.toto;
 
 class A {
-	public int method1(int i, int i) {
+	public int method1(int i, float i) {
 		return 1;
 	}
 }
